tergell_work/requirements.txt
apify-client
python-dotenv
scikit-learn
numpy
google-generativeai